% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppMirai-package.R
\docType{package}
\name{RcppMirai-package}
\alias{RcppMirai}
\alias{RcppMirai-package}
\title{RcppMirai: Calling an Rcpp function with mirai}
\description{
Provides an example implementation using a package to house a C++ function that is distributed in parallel across 'mirai' daemons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless-rd-rcpp/rcpp-and-mirai}
  \item \url{https://rd-rcpp.thecoatlessprofessor.com/rcpp-and-mirai/}
  \item Report bugs at \url{https://github.com/coatless-rd-rcpp/rcpp-and-mirai/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

Authors:
\itemize{
  \item James Balamuta \email{balamut2@illinois.edu}
}

}
\keyword{internal}
